Move BaseInvoiceTestCase
